Made server.js more modular Added 4 *route.js files in /server/ Currently rendering on 4000 broken as is: 3000/api/billing, and 3000/api/network api/cpu also has an undefined res may have to add instnace metrics to cpuRoutes.js to allow cpu querying
